requirements.txt
asgiref==3.6.0
certifi==2024.2.2
distlib==0.3.8
Django==4.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
filelock==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
pipenv==2023.12.1
platformdirs==4.2.2
setuptools==69.5.1
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.26.2
whitenoise==6.4.0
